feat: default 'Single Save File' option to false
The 'Single Save File' option allows stronger enforcement against save-scumming by default, but also presents the user with the possibility of an unrecoverable corrupted save if ONI burps out a bad save file at any point in the colony's history. It's safer to set this to false and rely on the non-invasive 'Hide Previous Saves' option, which still prevents the user from loading into a previous save game, but does it without mandating a single save file.